// src/types/app.ts
export enum AppActions {
    REQUEST_WEBAPP = "REQUEST_WEBAPP",
    RECEIVE_WEBAPP = "RECEIVE_WEBAPP",
    REQUEST_MODULES = "REQUEST_MODULES",
    RECEIVE_MODULES = "RECEIVE_MODULES",
    REQUEST_DELETE_MODULES = "REQUEST_DELETE_MODULES",
    RECEIVE_DELETE_MODULES = "RECEIVE_DELETE_MODULES",
    START_MODULE_UPDATE = "START_MODULE_UPDATE",
    END_MODULE_UPDATE = "END_MODULE_UPDATE",
    REQUEST_SAVE = "REQUEST_SAVE",
    RECEIVE_SAVE = "RECEIVE_SAVE",
    REQUEST_CREATE = "REQUEST_CREATE",
    RECEIVE_CREATE = "RECEIVE_CREATE",
    REQUEST_DELETE = "REQUEST_DELETE",
    RECEIVE_DELETE = "RECEIVE_DELETE",
    UPDATE_FILE_STATE = "UPDATE_FILE_STATE",
    UPDATE_APP_DATA = "UPDATE_APP_DATA",
    START_CLONING = "START_CLONING",
    END_CLONING = "END_CLONING"
}